+++ src/app/types/discord/DiscordUserCollection.ts
@@ -7,6 +7,9 @@ export interface DiscordUserCollection extends Collection {
 	_id: ObjectId,
 	userId: string,
 	tag: string,
-	isDMEnabled: boolean,
 	reportedForPOAP: number,
+	ethWalletSettings: {
+		publicAddress: string,
+		isPOAPDeliveryEnabled: boolean,
+	}
 }