+++ CHANGELOG.md
@@ -8,7 +8,8 @@
    - add ethWalletSettings to discordUsers
    - disable DM delivery
 4. Update poap expiration only for dates longer than 24 hours
-5. Create a private channel for in-channel flow
+5. Fix duplicate poap assignment
+6. Create a private channel for in-channel flow
 
 ## 2.7.0-RELEASE (2022-01-18)